move UpdateKeyRequest to web - wasn't caught as rename since diff is too different
