+++ apps/web/src/app/auth/key-rotation/request/update-key.request.ts
@@ -1,0 +1,18 @@
+import { OrganizationUserResetPasswordWithIdRequest } from "@bitwarden/common/admin-console/abstractions/organization-user/requests";
+import { SendWithIdRequest } from "@bitwarden/common/src/tools/send/models/request/send-with-id.request";
+import { CipherWithIdRequest } from "@bitwarden/common/src/vault/models/request/cipher-with-id.request";
+import { FolderWithIdRequest } from "@bitwarden/common/src/vault/models/request/folder-with-id.request";
+
+import { EmergencyAccessWithIdRequest } from "../../emergency-access/request/emergency-access-update.request";
+
+
+export class UpdateKeyRequest {
+  masterPasswordHash: string;
+  key: string;
+  privateKey: string;
+  ciphers: CipherWithIdRequest[] = [];
+  folders: FolderWithIdRequest[] = [];
+  sends: SendWithIdRequest[] = [];
+  emergencyAccessKeys: EmergencyAccessWithIdRequest[] = [];
+  resetPasswordKeys: OrganizationUserResetPasswordWithIdRequest[] = [];
+}
